Initial version
This provides the initial version which provides an upload form and
overview per day; repository; user; and worflow.

Basic styling is added to show the tables in an accounting style which
makes viewing it easier (and some PHP code to guess the needed number of
decimals). As well as some JavaScript so you can show/hide tables as
needed (e.g. showing "per day" for the 180 day overview can be
overwhelming).